+++ cloud/blockstore/tools/disaster_recovery/merge_path_description_backups/options.cpp
@@ -22,15 +22,10 @@ TOptions::TOptions(int argc, char** argv)
         .RequiredArgument("STR")
         .StoreResult(&BinaryOutputPath);
 
-    opts.AddLongOption("ss-file-name")
+    opts.AddLongOption("path-description-file-name")
         .RequiredArgument("STR")
         .DefaultValue("nbs-path-description-backup.txt")
         .StoreResult(&SchemeShardBackup);
-
-    opts.AddLongOption("hive-file-name")
-        .RequiredArgument("STR")
-        .DefaultValue("nbs-tablet-boot-info-backup.txt")
-        .StoreResult(&HiveBackup);
 
     TOptsParseResultException(&opts, argc, argv);
 }
