+++ cloud/blockstore/tools/disaster_recovery/merge_path_description_backups/ya.make
@@ -7,7 +7,6 @@ SRCS(
 
 PEERDIR(
     cloud/storage/core/libs/common
-    cloud/storage/core/libs/hive_proxy/protos
     cloud/storage/core/libs/ss_proxy/protos
 
     library/cpp/getopt
