+++ cloud/blockstore/tools/disaster_recovery/ya.make
@@ -1,0 +1,3 @@
+RECURSE(
+    merge_path_description_backups
+)
