--- cloud/blockstore/tools/restore/ya.make
@@ -1,3 +1,0 @@
-RECURSE(
-    merge_ss_backups
-)
